Separate library and binary into separate crates
Library is isolated now and we have two workspace members: vexel
and vexel-cli. This allows to isolate deps a bit, but the problem
with writer still exists. We need image and webp dependencies for
library tests and for binary, so library builds these dependencies
even thought they are only needed for tests. This is a problem for
later.